#108: moving CDM full & lite to new folder
