+++ PYTHON_CDM_Conversion_code/full_cdm_conversion/hourly_qff_to_cdm_lite_v1.py
