--- PYTHON_CDM_Conversion_code/quality_control_tables_details/test
@@ -1,1 +1,0 @@
-
