--- PYTHON_CDM_Conversion_code/run_diff_process
@@ -1,4 +1,0 @@
-#!/bin/bash
-source /apps/jasmin/jaspy/miniconda_envs/jaspy3.7/m3-4.6.14/bin/activate jaspy3.7-m3-4.6.14-r20200606
-#
-python  /gws/nopw/j04/c3s311a_lot2/data/level2/land/r202208/daily_updates/code/daily_diff_insert_to_cdm_jas_202204.py
