--- PYTHON_CDM_Conversion_code/source_files_format_details/test
@@ -1,1 +1,0 @@
-
